refactor:优化了utils文件夹的目录结构、代码文件
