+++ hkbbs-web/src/utils/api/helper/req.ts
@@ -1,0 +1,6 @@
+import Request from '../request'
+
+const BASE_URL = ""
+const TIMEOUT = 10000
+
+export const req = new Request(BASE_URL, TIMEOUT)
