+++ hkbbs-web/src/utils/api/index.ts
@@ -1,0 +1,8 @@
+import { Card } from "./helper/card"
+import { Hole } from "./helper/hole"
+import { Post } from "./helper/post"
+import { Both } from "./helper/both"
+import { Notice } from "./helper/notice"
+import { User } from "./helper/user"
+
+export default { Card, Hole, Post, Both, Notice, User }
